requirements.txt
aiogram
python-ldap